Updating PosteriorAnalysis to save cs and full mrtidal curve for each EOS
